% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionrate.R
\name{DetectionRate}
\alias{DetectionRate}
\alias{DetectionRate.Seurat}
\title{DetectionRate}
\usage{
DetectionRate(object, ...)

\method{DetectionRate}{Seurat}(object, features = NULL, slot_use = "data", thresh.min = 0, ...)
}
\arguments{
\item{object}{Seurat object}

\item{...}{ignored}

\item{features}{Which features to calculate detection rate for. Default: NULL (= all)}

\item{slot_use}{Slot to pull data from.  Default: "data"}

\item{thresh.min}{Minimum threshold to define 'detected' (log-scale)}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.


}
\description{
DetectionRate
}
