% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{GObubbleplot}
\alias{GObubbleplot}
\title{GObubbleplot}
\usage{
GObubbleplot(seuratObj, go_term, group.by = "ident", filter = NULL,
  do.return = FALSE, ...)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{go_term}{Gene Ontology term identifier (i.e. GO:0046774)}

\item{group.by}{Factor by which to group cells.  (default: ident)}

\item{filter}{A list of gene names to filter the GO term members against. (default: all genes in seuratObj)}

\item{do.return}{If TRUE, return a ggplot2 object instead of displaying chart}
}
\value{
if isTRUE(do.return), a ggplot2 object
}
\description{
Produces a Bubble Plot for the genes of a given GO term.
}
\examples{
GObubbleplot(seuratObj = dataset, go_term = "GO:0002253", filter = dataset@var.genes)
}
