% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_color_scale.R
\name{make_color_scale}
\alias{make_color_scale}
\title{make_color_scale}
\usage{
make_color_scale(palette = "viridis", gradations = 10)
}
\arguments{
\item{palette}{One of the palettes in RColorBrewer, viridis, or a list of colors understood by colors()}

\item{gradations}{The number of colors to return}
}
\value{
A list of hexidecimal numbers.
}
\description{
Using a list of colors, create a gradient of a discrete number of colors.
}
\examples{
make_color_scale(palette = 'Reds', n = 100)
}
