% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\alias{bubbleplot.Seurat}
\title{bubbleplot}
\usage{
bubbleplot(object, ...)

\method{bubbleplot}{Seurat}(
  object,
  features_plot,
  assay = NULL,
  slot = "data",
  filter_exp_pct = NULL,
  filter_apply_method = "any",
  filter_apply_group = NULL,
  filter_exp_pct_thresh = 0,
  filter_exp_level = 0,
  avg_func = "mean",
  grouping_var = "ident",
  x_lab_size = 9,
  y_lab_size = 9,
  x_axis_title = "Features",
  y_axis_title = "Grouping",
  pct_legend_title = "Percent group expressing",
  scale_legend_title = "Average scaled expression",
  x_lab_rot_angle = 45,
  preserve_feature_order = FALSE,
  cluster_x = TRUE,
  cluster_y = FALSE,
  colors_use = NULL,
  translate_feature_names = FALSE,
  annotated_feature_list = FALSE,
  do_return = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{...}{ignored}

\item{features_plot}{Either a list of features or a data frame of annotated features
to display (see 'annotated_feature_list'). Note: Gene and protein names
may be converted to the proper feature name automagically by
HGNChelper::checkGeneSymbols if 'translate_feature_names' is TRUE.
Features not appearing in the dataset are skipped.}

\item{assay}{Assay to plot. Default: "RNA"}

\item{slot}{Slot to plot. Default: "data"}

\item{filter_exp_pct}{Display only features that are expressed above this
fraction.  The method by which and groups in which this is judged are controlled by
`filter_apply_method` and `filter_apply_group`, respectively.  Default: NULL}

\item{filter_apply_method}{How should `filter_exp_pct` be applied - should `any` or `all` groups
be above the desired expression level? Default: "any"}

\item{filter_apply_group}{What identity groups should be considered when applying 
`filter_exp_pct?  Only works with the active identity or whatever is specified in `grouping_var`. 
Default: NULL}

\item{filter_exp_pct_thresh}{Threshold for expression fraction. Default: 0}

\item{filter_exp_level}{Display only features that are expressed above this
level in at least one group. Default: 0}

\item{avg_func}{Base function to call for measuring average expression. Default: "mean"}

\item{grouping_var}{Variable by which to group cells.  Can be cluster identities
or any column from the meta.data slot. Default: ident}

\item{x_lab_size}{Font size for the x-axis labels. Default: 9}

\item{y_lab_size}{Font size for the y-axis labels. Default: 9}

\item{x_axis_title}{x-axis title}

\item{y_axis_title}{y-axis title}

\item{pct_legend_title}{Percent expressed legend title}

\item{scale_legend_title}{Scale legend title.}

\item{x_lab_rot_angle}{Angle to rotate the x-axis labels. Default: 45°}

\item{preserve_feature_order}{Should the features by displayed in order in which
they are given? Overrides cluster_x. Default: FALSE.}

\item{cluster_x}{Arrange the x-axis variables using hierarchical clustering.
Default: TRUE}

\item{cluster_y}{Arrange the y-axis variables using hierarchical clustering.
Default: FALSE}

\item{colors_use}{Color palette to use to display expression levels.
Default: "Reds"}

\item{translate_feature_names}{Should feature names be checked and translated to
their correct HUGO Gene Nomenclature Committee name? Default: FALSE}

\item{annotated_feature_list}{Is the feature list annotated?  If so, the features will be
faceted using their annontations.  Requires that 'features_plot' is a two column
with the annotations in a column named 'annotations' and features in a column
named 'features'. Default: FALSE}

\item{do_return}{Return a ggplot2 object instead of displaying}

\item{verbose}{Show extra output information, like the features that were not 
found? Default: FALSE}
}
\value{
if isTRUE(do_return), a ggplot2 object


}
\description{
Display percentage of cells expressing and levels for a set of features.
}
