% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\alias{bubbleplot.seurat}
\alias{bubbleplot.Seurat}
\title{bubbleplot}
\usage{
bubbleplot(object, ...)

\method{bubbleplot}{seurat}(object, genes_plot, use_scaled = FALSE,
  filter_exp_pct = NULL, filter_exp_pct_thresh = 0,
  filter_exp_level = 0, group_by = "ident", x_lab_size = 9,
  y_lab_size = 9, x_axis_title = "Genes", y_axis_title = "Grouping",
  pct_legend_title = "Percent group expressing",
  scale_legend_title = "Average scaled expression",
  x_lab_rot_angle = 45, preserve_gene_order = FALSE,
  cluster_x = TRUE, cluster_y = FALSE, colors_use = NULL,
  translate_gene_names = FALSE, annotated_gene_list = FALSE,
  do_return = FALSE, verbose = FALSE)

\method{bubbleplot}{Seurat}(object, genes_plot, assay = "RNA",
  slot = "data", filter_exp_pct = NULL, filter_exp_pct_thresh = 0,
  filter_exp_level = 0, group_by = "ident", x_lab_size = 9,
  y_lab_size = 9, x_axis_title = "Genes", y_axis_title = "Grouping",
  pct_legend_title = "Percent group expressing",
  scale_legend_title = "Average scaled expression",
  x_lab_rot_angle = 45, preserve_gene_order = FALSE,
  cluster_x = TRUE, cluster_y = FALSE, colors_use = NULL,
  translate_gene_names = FALSE, annotated_gene_list = FALSE,
  do_return = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{genes_plot}{Either a list of genes or a data frame of annotated genes
to display (see 'annotated_gene_list'). Note: Gene and protein names
may be converted to the proper gene name automagically by
HGNChelper::checkGeneSymbols if 'translate_gene_names' is TRUE.
Genes not appearing in the dataset are skipped.}

\item{filter_exp_pct}{Display only genes that are expressed above this
fraction of cells in at least one group. Default: NULL}

\item{filter_exp_pct_thresh}{Threshold for expression fraction. Default: 0}

\item{filter_exp_level}{Display only genes that are expressed above this
level in at least one group. Default: 0}

\item{group_by}{Variable by which to group cells.  Can be cluster identities
or any column from the meta.data slot. Default: ident}

\item{x_lab_size}{Font size for the x-axis labels. Default: 9}

\item{y_lab_size}{Font size for the y-axis labels. Default: 9}

\item{x_lab_rot_angle}{Angle to rotate the x-axis labels. Default: 45°}

\item{preserve_gene_order}{Should the genes by displayed in order in which
they are given? Default: FALSE.}

\item{cluster_x}{Arrange the x-axis variables using hierarchical clustering.
Default: TRUE}

\item{cluster_y}{Arrange the y-axis variables using hierarchical clustering.
Default: FALSE}

\item{colors_use}{Color palette to use to display expression levels.
Default: "Reds"}

\item{translate_gene_names}{Should gene names be checked and translated to
their correct HUGO Gene Nomenclature Committee name? Default: FALSE}

\item{annotated_gene_list}{Is the gene list annotated?  If so, the genes will be
faceted using their annontations.  Requires that 'genes_plot' is a two column
with the annotations in a column named 'annotations' and genes in a column
named 'genes'. Default: FALSE}

\item{do_return}{Return a ggplot2 object instead of displaying}

\item{verbose}{Show extra output information, like the genes that were not 
found? Default: FALSE}

\item{assay}{Assay to plot. Default: "RNA"}

\item{slot}{Slot to plot. Default: "data"}
}
\value{
if isTRUE(do_return), a ggplot2 object
}
\description{
Display percentage of cells expressing and levels for a set of genes.
}
\examples{
BubblePlot(object = obj, genes_plot = c("IFIT1","IFITM1","IFITM3"), group_by = "treatment")
}
