% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctgenenames.R
\name{correctGeneNames}
\alias{correctGeneNames}
\title{correctGeneName}
\usage{
correctGeneNames(seuratObj)
}
\arguments{
\item{seuratObj}{}
}
\value{
seurat object
}
\description{
Convert the gene names in a Seurat object to the accepted HUGO gene names
}
