% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctgenenames.R
\name{correctGeneNames}
\alias{correctGeneNames}
\alias{correctGeneNames.Seurat}
\title{correctGeneNames}
\usage{
correctGeneNames(object, ...)

\method{correctGeneNames}{Seurat}(object, assay = "RNA")
}
\arguments{
\item{object}{}
}
\value{
seurat object
}
\description{
Convert the gene names in a Seurat object to the accepted HUGO gene names
}
