% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averagedetectionrate.R
\name{DetectionRate}
\alias{DetectionRate}
\alias{DetectionRate.Seurat}
\title{DetectionRate}
\usage{
DetectionRate(object, ...)

\method{DetectionRate}{Seurat}(object, assay = "RNA", features = NULL,
  slot_use = "data", thresh.min = 0, ...)
}
\arguments{
\item{object}{Seurat object}

\item{thresh.min}{Minimum threshold to define 'detected' (log-scale)}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
}
\description{
DetectionRate
}
\examples{
head(DetectionRate(object = pbmc_small))
}
