% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{GObubbleplot}
\alias{GObubbleplot}
\title{GObubbleplot}
\usage{
GObubbleplot(seuratObj, go_term, group.by = "ident", filter = NULL,
  filter.exp.pct = NULL, filter.exp.pct.thresh = 0,
  filter.exp.level = 0, x.lab.size = 9, y.lab.size = 9,
  x.lab.rot.angle = 45, cluster.x = TRUE, cluster.y = TRUE,
  colors.use = NULL, do.return = FALSE, ...)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{go_term}{Gene Ontology term identifier (i.e. GO:0046774)}

\item{group.by}{Factor by which to group cells.  (default: ident)}

\item{filter}{A list of gene names to filter the GO term members against. (default: all genes in seuratObj)}

\item{filter.exp.pct}{Display only genes that are expressed above this fraction of cells in at least one group. (default: NULL)}

\item{filter.exp.pct.thresh}{Threshold for expression fraction. (default: 0)}

\item{filter.exp.level}{Display only genes that are expressed above this level in at least one group. (default: 0)}

\item{do.return}{If TRUE, return a ggplot2 object instead of displaying chart}

\item{...}{Additional options to pass to bubbleplot}
}
\value{
if isTRUE(do.return), a ggplot2 object
}
\description{
Produces a Bubble Plot for the genes of a given GO term.
}
\examples{
GObubbleplot(seuratObj = dataset, go_term = "GO:0002253", filter = dataset@var.genes)
}
